Making the side info bar a fixed size on greater 992 px
previously using bootstrap the sidebar was split 3/ 9 with the map, but it seemed unecessary to have so much space for the side bar especially as the screen got wider. I used flex box and flex:1 and also tried using it in more places.

cleaned up css
and messed more with spacing in the sidebar