remove bootstraps grid styles
used flexbox on the thumbnail to be better able control the max width of the image
for the header, the bootstrap grid styles where also removed, 320px wide the h1 clashed with menu button even though there was more space. postion absolute was used.

the pageHeader class that added a underline was removed and a breakline was used instead since it spans the entire width of the page

css file was cleaned up on more of a logical order